requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
google-generativeai==0.8.2
python-dotenv==0.19.2
